Major changes to the Python module. (This module will be archived very soon.)

+++ uninstall.sh
@@ -1,5 +1,12 @@
 #!/usr/bin/bash
 
-sudo rm rf - /usr/bin/calc
+read -p "Do you want to remove the Calc application by deleting it from the /usr/bin directory? " answer
+
+if [[ "$answer" == "y" || "$answer" == "Y" ]]; then
+    sudo rm rf - /usr/bin/calc
+    echo "The Calc app has been successfully uninstalled."
+else
+    echo "The uninstallation of the Calc app has been canceled."
+fi
 
 # Changes will be made to the uninstall files for Linux.
